Fix int32_t overflow in av1_round_shift_array_c
BUG=aomedia:1960

+++ av1/common/av1_txfm.c
@@ -78,7 +78,7 @@ void av1_round_shift_array_c(int32_t *arr, int size, int bit) {
       }
     } else {
       for (i = 0; i < size; i++) {
-        arr[i] = arr[i] * (1 << (-bit));
+        arr[i] = (int32_t)((int64_t)arr[i] * (1ULL << (-bit)));
       }
     }
   }
